tweak timeouts, rename to setupAcceptance
